pyproject.toml
[project]
name = "handmade-rag"
version = "0.1.0"
description = "本地端 RAG 問答系統，針對 GIGABYTE AORUS MASTER 16 AM6H 系列筆電規格設計"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "llama-cpp-python>=0.3.16",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "requests>=2.32.5",
]
